Sprint 22

+++ app/src/main/java/data/player/FavoriteRepository.kt
@@ -7,7 +7,7 @@ interface FavoriteRepository {
 
     suspend fun add(track: Track)
 
-    suspend fun getById(id: Int): Track?
+    suspend fun getById(id: Long): Track?
 
     suspend fun delete(track: Track)
