+++ app/src/main/java/domain/player/FavoriteInteractorImpl.kt
@@ -9,7 +9,7 @@ class FavoriteInteractorImpl(
 
     override suspend fun add(track: Track) = repository.add(track)
 
-    override suspend fun getById(id: Int): Track? = repository.getById(id)
+    override suspend fun getById(id: Long): Track? = repository.getById(id)
 
     override suspend fun delete(track: Track) = repository.delete(track)
